Frenquencia por decada.

Retorna a frequencia de nascimentos por decada de nascimento para o nome consultado.

brnome_freq(nome, sexo = NULL, localidade_cod = NULL)

Arguments

nome
Vetor com um ou mais nomes a serem consultados.

sexo
Se NULL (default), a consulta pelo nome eh unissex. Valores podem ser "M", para o sexo masculino, ou "F", para o feminino.

localidade_cod
Se NULL (default), a consulta eh para todo territorio. Para especificar uma localidade, utilize o identificador de uma Unidade da Federação.

Examples

brnome_freq("italo", "m")

#> # A tibble: 9 x 7
#>   nome  sexo  localidade_cod localidade_nome nascimento_peri~ nascimento_deca~
#>   <chr> <chr> <lgl>          <chr>           <chr>                       <int>
#> 1 ITALO M     NA             Brasil          [    , 1930]                 1930
#> 2 ITALO M     NA             Brasil          [1930, 1940]                 1930
#> 3 ITALO M     NA             Brasil          [1940, 1950]                 1940
#> 4 ITALO M     NA             Brasil          [1950, 1960]                 1950
#> 5 ITALO M     NA             Brasil          [1960, 1970]                 1960
#> 6 ITALO M     NA             Brasil          [1970, 1980]                 1970
#> 7 ITALO M     NA             Brasil          [1980, 1990]                 1980
#> 8 ITALO M     NA             Brasil          [1990, 2000]                 1990
#> 9 ITALO M     NA             Brasil          [2000, 2010]                 2000
#> # ... with 1 more variable: frequencia <int>